refactor PlanTable component structure and styling; move to dedicated files
